feat: working segmentation highlight
